fix: make all syntax tree struct fields public
As a parse target, the parser module needs to be able
to instantiate the syntax tree types. Since they are
separated into different modules, either the syntax
tree types need to expose public constructor functions
or the fields need to be made public. I opted to make
the fields public since it is 1) less verbose 2) easier
to maintain and 3) the fields are still immutable so
we don't need to worry about other modules mistakenly
mutating the public state

Signed-off-by: Matthew Pickering <wyomatthew@hotmail.com>